+++ cloudflare/src/lib.rs
@@ -1,4 +1,4 @@
-use std::sync::Arc;
+use std::rc::Rc;
 
 use anyhow::anyhow;
 use tailcall::io::{EnvIO, FileIO, HttpIO};
@@ -9,11 +9,11 @@ mod handle;
 mod http;
 mod r2_address;
 
-pub fn init_env(env: Arc<worker::Env>) -> impl EnvIO {
+pub fn init_env(env: Rc<worker::Env>) -> impl EnvIO {
   env::EnvCloudflare::init(env)
 }
 
-pub fn init_file(env: Arc<worker::Env>) -> impl FileIO {
+pub fn init_file(env: Rc<worker::Env>) -> impl FileIO {
   file::CloudflareFileIO::init(env)
 }
 
@@ -22,22 +22,8 @@ pub fn init_http() -> impl HttpIO + Default + Clone {
 }
 
 #[worker::event(fetch)]
-async fn fetch(req: worker::Request, env: worker::Env, context: worker::Context) -> anyhow::Result<worker::Response> {
-  let result = handle::fetch(req, env, context).await;
-
-  match result {
-    Ok(response) => Ok(response),
-    Err(message) => {
-      log::error!("ServerError: {}", message.to_string());
-      worker::Response::error(message.to_string(), 500).map_err(to_anyhow)
-    }
-  }
-}
-
-#[worker::event(start)]
-fn start() {
-  // Initialize Logger
-  wasm_logger::init(wasm_logger::Config::new(log::Level::Info));
+async fn main(req: worker::Request, env: worker::Env, context: worker::Context) -> anyhow::Result<worker::Response> {
+  Ok(handle::execute(req, env, context).await?)
 }
 
 fn to_anyhow<T: std::fmt::Display>(e: T) -> anyhow::Error {
