+++ src/cli/file.rs
@@ -11,7 +11,6 @@ impl NativeFileIO {
   }
 }
 
-#[async_trait::async_trait]
 impl FileIO for NativeFileIO {
   async fn write<'a>(&'a self, file: &'a str, content: &'a [u8]) -> Result<()> {
     let mut file = tokio::fs::File::create(file).await?;
