+++ src/io.rs
@@ -1,4 +1,5 @@
 use std::collections::HashMap;
+use std::future::Future;
 
 use crate::http::Response;
 pub trait EnvIO: Send + Sync {
@@ -13,11 +14,10 @@ pub trait HttpIO: Sync + Send {
   async fn execute_raw(&self, request: reqwest::Request) -> anyhow::Result<Response<Vec<u8>>>;
 }
 
-#[async_trait::async_trait]
 pub trait FileIO {
-  async fn write<'a>(&'a self, file: &'a str, content: &'a [u8]) -> anyhow::Result<()>;
-  async fn read<'a>(&'a self, file_path: &'a str) -> anyhow::Result<String>;
-  async fn read_all<'a>(&'a self, file_paths: &'a [String]) -> anyhow::Result<Vec<(String, String)>>;
+  fn write<'a>(&'a self, file: &'a str, content: &'a [u8]) -> impl Future<Output = anyhow::Result<()>>;
+  fn read<'a>(&'a self, file_path: &'a str) -> impl Future<Output = anyhow::Result<String>>;
+  fn read_all<'a>(&'a self, file_paths: &'a [String]) -> impl Future<Output = anyhow::Result<Vec<(String, String)>>>;
 }
 
 // TODO: rename to ConstEnv
