+++ tests/graphql_spec.rs
@@ -7,6 +7,7 @@ use std::sync::{Arc, Once};
 use async_graphql::parser::types::TypeSystemDefinition;
 use async_graphql::Request;
 use derive_setters::Setters;
+use futures_util::future::join_all;
 use hyper::http::{HeaderName, HeaderValue};
 use hyper::HeaderMap;
 use pretty_assertions::{assert_eq, assert_ne};
@@ -325,9 +326,7 @@ async fn test_execution() -> std::io::Result<()> {
     })
     .collect();
 
-  for task in tasks {
-    task.await?;
-  }
+  join_all(tasks).await;
 
   Ok(())
 }
